[refactor][m]: refactor components using reusable component
