+++ components/dataset/Resources.tsx
@@ -1,38 +1,41 @@
 import Link from 'next/link';
 import { useQuery } from '@apollo/react-hooks';
-import ErrorMessage from '../Error';
+import { Table, ErrorMessage } from '../_shared';
 import { GET_RESOURCES_QUERY } from '../../graphql/queries';
 
 const columns = [
   {
     name: 'File',
-    render: ({ name: resName, title }, { name }) => (
-      <Link href={`${name}/r/${resName}`}>
+    key: 'file',
+    render: ({ name: resName, title, parentName }) => (
+      <Link href={`${parentName}/r/${resName}`}>
         <a className="underline">{title || resName}</a>
       </Link>
     ),
   },
   {
     name: 'Format',
-    render: ({ format }) => format,
+    key: 'format',
   },
   {
     name: 'Created',
-    render: ({ created }) => created,
+    key: 'created',
   },
   {
     name: 'Updated',
-    render: ({ last_modified }) => last_modified,
+    key: 'last_modified',
   },
   {
     name: 'Link',
-    render: ({ name }, { name: resName }) => (
-      <Link href={`${name}/r/${resName}`}>
+    key: 'link',
+    render: ({ name: resName, parentName }) => (
+      <Link href={`${parentName}/r/${resName}`}>
         <a className="underline">Preview</a>
       </Link>
     ),
   },
 ];
+
 export default function Resources({ variables }) {
   const { loading, error, data } = useQuery(GET_RESOURCES_QUERY, {
     variables,
@@ -50,27 +53,13 @@ export default function Resources({ variables }) {
   return (
     <>
       <h3 className="text-xl font-semibold">Data Files</h3>
-      <table className="table-auto w-full text-sm text-left mb-6">
-        <thead>
-          <tr>
-            {columns.map(({ name }) => (
-              <th className="px-4 py-2">{name}</th>
-            ))}
-          </tr>
-        </thead>
-        <tbody>
-          {/* Do not use array index as key */}
-          {result.resources.map((resource) => (
-            <tr key={resource.id}>
-              {columns.map(({ name, render }) => (
-                <td key={name} className="px-4 py-2">
-                  {render(resource, result)}
-                </td>
-              ))}
-            </tr>
-          ))}
-        </tbody>
-      </table>
+      <Table
+        columns={columns}
+        data={result.resources.map((resource) => ({
+          ...resource,
+          parentName: result.name,
+        }))}
+      />
     </>
   );
 }
